website/requirements.txt
dj-database-url
Django
django-heroku
gunicorn
psycopg2
pytz
sqlparse
whitenoise
farm-haystack==1.8.0
duckduckgo-search==2.1.3
pandas==1.1.0